Re-organise the projects
