+++ projects/ideas.md
@@ -1,4 +1,10 @@
 # Project Ideas
 
-* k-means image segmentation
-* Compare data compression methods
+* k-means image segmentation to identify cells from flourencence microscopy images
+* Compare data compression methods on sequence data
+  - RLE
+  - Dictionary
+  - Huffman codes
+* Decode a BMP image (maybe chest xray)
+  - Ref: https://www.kaggle.com/paultimothymooney/chest-xray-pneumonia
+* Plot a 3D protein structure by reading PDB data
